requirements_erosion.txt
# Requirements for Terrain Erosion Prediction
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
xgboost>=1.5.0
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-image>=0.18.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
joblib>=1.0.0

